Re-factored tournament.model.ts to become tournament.interfaces.ts, added the create and patch interfaces
