+++ src/model/tournament/tournament.interfaces.ts
@@ -1,0 +1,59 @@
+import { keys } from 'ts-transformer-keys';
+/**
+ *
+ */
+export interface TournamentDto {
+	id: string;				// created by the server
+	name: string;			// must be unique
+	city?: string;
+	country?: string;
+	month?: number;
+	year?: number;
+	rounds: number;
+	maxPlayers: number;
+	type: string; //roundRobin, swiss, elimination, match
+	players?: Array<number> // TournamentPlayerDto.id
+	state?: string; // planned* / scheduled / closed / underway / complete
+	winPoints?: number; // points awarded for a win, 1 is default
+	tiePoints?: number; // points awarded for a tie, 0.5 is default
+	started?: number; // milliseconds since epoch to date (date when transitioned to underway)
+	ended?: number; // milliseconds since epoch to date (date when transitioned to complete)
+}
+export const TOURNAMENT_DTO_KEYS = keys<TournamentDto>();
+
+/**
+ *  Interface used to create an user entity. Note that the remaining UserDto attributes are generated by the service
+ *  when creating an entity;
+ *
+ *  The CREATE API returns a UserDto for the new entity;
+ */
+export interface ITournamentCreate {
+	name: string;			// must be unique
+	rounds: number;
+	maxPlayers: number;
+	type: string; //roundRobin, swiss, elimination, match
+}
+export const USER_CREATE_KEYS = keys<ITournamentCreate>();
+/**
+ *  Interface used to create a user entity. Note that the remaining attributes are generated by the service when
+ *  patching an entity;
+ *
+ *  The PATCH API returns a UserDto for the updated entity;
+ */
+export interface ITournamentPatch {
+	name?: string;			// must be unique
+	city?: string;
+	country?: string;
+	month?: number;
+	year?: number;
+	rounds?: number;
+	maxPlayers?: number;
+	type?: string; //roundRobin, swiss, elimination, match
+	players?: Array<number> // TournamentPlayerDto.id
+	state?: string; // planned* / scheduled / closed / underway / complete
+	winPoints?: number; // points awarded for a win, 1 is default
+	tiePoints?: number; // points awarded for a tie, 0.5 is default
+	started?: number; // milliseconds since epoch to date (date when transitioned to underway)
+	ended?: number; // milliseconds since epoch to date (date when transitioned to complete)
+}
+export const USER_PATCH_KEYS = keys<ITournamentPatch>();
